Lost all hope. Moved all of the operators code to Eval.Eval
